moving some core libs around
