+++ elements/bulbs-poll/fields/vote.js
@@ -1,3 +1,4 @@
+import makeRequest from 'bulbs-elements/util/make-request';
 import { Field, Action } from 'bulbs-elements/store';
 
 function cacheKey (pollId) {
@@ -18,7 +19,8 @@ const VoteField = new Field({
   }),
   makeVoteRequest: new Action(function (state, answer, store) {
     let { poll } = store.state;
-    this.request(`https://onion.sodahead.com/api/polls/${poll.data.sodahead_id}/vote/`, {
+    let url = `https://onion.sodahead.com/api/polls/${poll.data.sodahead_id}/vote/`
+    makeRequest(url, {
       method: 'post',
       headers: {
         'Accept': 'application/json',
