+++ elements/campaign-display/campaign-display-store.js
@@ -2,6 +2,8 @@ import { Store } from 'bulbs-elements/store';
 
 export default class CampaignDisplayStore extends Store {}
 
+import CampaignField from './fields/campaign';
+
 Store.defineFields(Store, {
-  campaign: CampaignDisplayStore,
+  campaign: CampaignField,
 });
