+++ lib/bulbs-elements/store/action.js
@@ -1,0 +1,5 @@
+export default class Action {
+  constructor (handler) {
+    this.invoke = handler;
+  }
+}
