+++ lib/bulbs-elements/store/index.js
@@ -1,0 +1,5 @@
+import Store from './store';
+import Action from './action';
+import Field from './field';
+
+module.exports = { Store, Action, Field };
