+++ package.json
@@ -89,6 +89,7 @@
     "document-register-element": "^0.5.4",
     "dom4": "^1.6.0",
     "es6-promise": "^3.0.2",
+    "invariant": "^2.2.1",
     "isomorphic-fetch": "^2.2.1",
     "object-map-to-array": "^1.0.0",
     "react": "^0.14.6",
